feat(release): v2.1.1 (82) Play Store closed testing release, visual screen navigation guide, and UI polish

+++ app/src/main/java/com/bit/ui/components/M3ProgressIndicators.kt
@@ -292,7 +292,7 @@ fun M3WavyLinearProgressIndicator(
             wavePath.moveTo(0f, centerY + ampPx * sin((phase).toDouble()).toFloat())
 
             while (x <= activeWidth) {
-                val y = centerY + ampPx * sin(((x / waveLenPx) * 2 * PI + phase).toDouble()).toFloat()
+                val y = centerY + ampPx * sin((x / waveLenPx) * 2 * PI + phase).toFloat()
                 wavePath.lineTo(x, y)
                 x += step
             }
